requirements.txt
# install_requires
aiohttp
oauthlib
python-magic
